@@ -1,48 +1,48 @@
 {
-    "name": "ep-system-frontend",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve --port=4100",
-        "build": "ng build",
-        "watch": "ng build --watch --configuration development",
-        "test": "ng test"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "~13.2.0",
-        "@angular/cdk": "^13.3.9",
-        "@angular/common": "~13.2.0",
-        "@angular/compiler": "~13.2.0",
-        "@angular/core": "~13.2.0",
-        "@angular/fire": "^7.5.0",
-        "@angular/forms": "~13.2.0",
-        "@angular/material": "^13.3.9",
-        "@angular/material-moment-adapter": "13.3",
-        "@angular/platform-browser": "~13.2.0",
-        "@angular/platform-browser-dynamic": "~13.2.0",
-        "@angular/router": "~13.2.0",
-        "bootstrap": "^5.2.3",
-        "ngx-spinner": "^13.1.1",
-        "rxjs": "~7.5.0",
-        "tslib": "^2.3.0",
-        "uuid": "^9.0.1",
-        "zone.js": "~0.11.4"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~13.2.0",
-        "@angular/cli": "~13.2.0",
-        "@angular/compiler-cli": "~13.2.0",
-        "@ngxs/devtools-plugin": "^3.8.1",
-        "@types/jasmine": "~3.10.0",
-        "@types/node": "^12.11.1",
-        "@types/uuid": "^9.0.4",
-        "jasmine-core": "~4.0.0",
-        "karma": "~6.3.0",
-        "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage": "~2.1.0",
-        "karma-jasmine": "~4.0.0",
-        "karma-jasmine-html-reporter": "~1.7.0",
-        "typescript": "~4.5.2"
-    }
+  "name": "ep-system-frontend",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --port=4100",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/fire": "^7.5.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^13.3.9",
+    "@angular/material-moment-adapter": "13.3",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "bootstrap": "^5.2.3",
+    "ngx-spinner": "^13.1.1",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
+    "uuid": "^9.0.1",
+    "zone.js": "~0.11.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@ngxs/devtools-plugin": "^3.8.1",
+    "@types/jasmine": "~3.10.0",
+    "@types/node": "^12.11.1",
+    "@types/uuid": "^9.0.4",
+    "jasmine-core": "~4.0.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.1.0",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.8.4"
+  }
 }
